🔀::Boardgame 로직 개선

+++ src/main/java/com/gapple/weeingback/domain/consulting/entity/ToConsultingResponse.java
@@ -10,7 +10,7 @@ public class ToConsultingResponse {
 
     private Long issuedAt;
 
-    private int classTime;
+    private Long classTime;
 
     private String description;
 }
